+++ src/App.jsx
@@ -1,34 +1,10 @@
-import {
-  Header,
-  Download,
-  FAQs,
-  Footer,
-  Hero,
-  Newsletter,
-  Platforms,
-  Profile,
-  Stream,
-} from './containers';
-import Background from './components/Background';
+import { Hero } from './components';
 
 const App = () => {
   return (
-    <div>
-      <Background>
-        <Header />
-        <Hero />
-      </Background>
-
-      <div>
-        <Platforms />
-        <Download />
-        <Stream />
-        <Profile />
-        <FAQs />
-        <Newsletter />
-        <Footer />
-      </div>
-    </div>
+    <>
+      <Hero />
+    </>
   );
 };
