+++ src/assets/index.js
@@ -1,15 +1,19 @@
-import netflix from './netflix-logo.svg';
-import netflixBg from './cover.svg';
-import netflixTv from './netflix-on-tv.gif';
-import netflixPhone from './mobile-netflix.gif';
-import netflixPlatform from './netflix-platform.gif';
-import netflixProfiles from './netflix-profiles.gif';
+// import netflix from './netflix-logo.svg';
+// import netflixBg from './cover.jpg';
+// import netflixTv from './netflix-on-tv.gif';
+// import netflixPhone from './mobile-netflix.gif';
+// import netflixPlatform from './netflix-platform.gif';
+// import netflixProfiles from './netflix-profiles.gif';
 
-export {
-  netflix,
-  netflixBg,
-  netflixTv,
-  netflixPhone,
-  netflixPlatform,
-  netflixProfiles,
-};
+// export {
+//   netflix,
+//   netflixBg,
+//   netflixTv,
+//   netflixPhone,
+//   netflixPlatform,
+//   netflixProfiles,
+// };
+
+import heroBG from '../assets/cover.png';
+
+export { heroBG };
