+++ src/components/index.js
@@ -1,4 +1,6 @@
-export { default as SelectionMenu } from './SelectionMenu';
-export { default as Button } from './Button';
-export { default as GetStarted } from './GetStarted';
-export { default as Accordion } from './Accordion';
+// export { default as SelectionMenu } from './SelectionMenu';
+// export { default as Button } from './Button';
+// export { default as GetStarted } from './GetStarted';
+// export { default as Accordion } from './Accordion';
+
+export { default as Hero } from './Hero';
