--- src/styles.js
@@ -1,14 +1,0 @@
-const styles = {
-  heading: 'font-lato text-5xl font-extrabold text-white',
-  paragraph: 'font-lato text-2xl text-white',
-
-  flexCenter: 'flex justify-center items-center',
-  flexRowBetween: 'flex flex-row justify-between items-center',
-  flexFooter: 'flex flex-row justify-between items-start',
-
-  transparentBg: 'bg-[rgba(0,0,0,0.8)]',
-  bgBlack: 'bg-[rgb(0,0,0)]',
-  collageBg: 'bg-collage bg-cover bg-center',
-};
-
-export default styles;
