+++ src/ui/index.js
@@ -1,0 +1,3 @@
+export { default as GetStarted } from './GetStarted';
+export { default as Button } from './Button';
+export { default as styles } from './styles';
